requirements.txt
numpy==1.26.3
scikit-learn==1.4.0
pydantic==1.10.2
geopy==2.4.1
mlflow[extras]
fastapi==0.89.1
uvicorn==0.22.0
python-dotenv==1.0.0
asyncpg==0.29.0
sqlalchemy[asyncio]